refactoring main.rs

+++ src/string/string_1020.rs
@@ -50,7 +50,7 @@ fn string_op_1020_c2f8(param_1: u16) -> *mut u8
   case 0x10:
   }
   pcVar1 = load_string_1010_847e
-                     (_PTR_LOOP_1050_14cc,(INT16)(_PTR_LOOP_1050_14cc >> 0x10)
+                     (_PTR_LOOP_1050_14cc,(_PTR_LOOP_1050_14cc >> 0x10)
                       ,0x1010);
   return pcVar1;
 }
